requirements.txt
# Turd Gambling Network Dependencies
discord.py>=2.6.0
